Use select for name of instituitions instead of input
