+++ src/components/SubmitForm.jsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { useState, useRef, useEffect } from "react";
+import nigerianUniversities from "@/utils/nigerianUniversities";
 
 export default function SubmitPastQuestionForm() {
   const [form, setForm] = useState({
@@ -304,13 +305,11 @@ export default function SubmitPastQuestionForm() {
               >
                 Name of Institution <span className="text-red-500">*</span>
               </label>
-              <input
+              <select
                 ref={formRef.nameOfInstitution}
                 id="name-of-institution"
                 name="name-of-institution"
-                type="text"
                 required
-                placeholder="e.g., Western Delta University, Oghara, Delta State"
                 value={form.nameOfInstitution}
                 onChange={(e) =>
                   setForm({
@@ -319,7 +318,19 @@ export default function SubmitPastQuestionForm() {
                   })
                 }
                 className="mt-2 block w-full rounded-md border-0 p-2 text-gray-900 shadow-sm ring-1 ring-inset ring-slate-300 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
+              >
+                <option value="">--Please choose an option--</option>
+                {Object.keys(nigerianUniversities).map((header) => (
+                  <optgroup key={header} label={header.toUpperCase()}>
+                    {nigerianUniversities[header].map((item, index) => (
+                      <option key={index} value={item?.name}>
+                        {item?.name}
+                      </option>
+                    ))}
+                  </optgroup>
+                ))}
+              </select>
+
               {formError.nameOfInstitution && (
                 <span className="text-xs text-red-500">
                   Name of institution cannot be empty
